GV-339 added the loader in the fund-details-statistics and in the program-details-statistics
